added mono specific type cast
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS05/]$/ZXingNet/trunk;C86516

+++ Source/lib/aztec/encoder/HighLevelEncoder.cs
@@ -184,7 +184,8 @@ namespace ZXing.Aztec.Internal
          for (int index = 0; index < text.Length; index++)
          {
             int pairCode;
-            int nextChar = index + 1 < text.Length ? text[index + 1] : 0;
+            // don't remove the (int) type cast, mono compiler needs it
+            int nextChar = (index + 1 < text.Length) ? (int)text[index + 1] : 0;
             switch (text[index])
             {
                case (byte)'\r':
